Cert/key related improvements to gate-multinode
* Adds new config option GATE_SSH_KEYS to enable user provided SSH key-pair.
  This applies to both virtmgr and access to virtual nodes

* Adds new config option USE_EXISTING_SECRETS to use already existing
  manifests for certificates and SSH key files

+++ tools/multi_nodes_gate/airship_gate/stages/generate-certificates.sh
@@ -30,6 +30,12 @@ GATE_FILES=($(find "${GATE_DEPOT}" -name '*.yaml' | xargs -n 1 basename | xargs 
 mkdir -p "${CERT_DEPOT}"
 chmod 777 "${CERT_DEPOT}"
 
+if [[ "${USE_EXISTING_SECRETS}" ]]
+then
+  log Certificates already provided by manifests
+  exit 0
+fi
+
 log Generating certificates
 docker run --rm -t \
     -w /tmp \
